Move loadConfig out into its own utility file
